requirements.txt
Orange3
orange3-network # to enable Gene Network widget
biopython
setuptools
numpy
scipy
six
genesis-pyapi>=1.1.2
slumber>=0.7.1
requests>=2.11.1
requests-cache>=0.4.12
serverfiles>=0.2
